# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: First-party Cognirise website and mini CMS API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/bootstrap:
    get:
      operationId: getAuthBootstrap
      tags: [auth]
      summary: Get initial account setup state
      responses:
        "200":
          description: Bootstrap state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthBootstrap"
    post:
      operationId: bootstrapAuth
      tags: [auth]
      summary: Create the initial administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthBootstrapInput"
      responses:
        "201":
          description: Administrator and session created
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login result or MFA challenge
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/mfa/verify:
    post:
      operationId: verifyMfa
      tags: [auth]
      summary: Complete an MFA challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaVerificationInput"
      responses:
        "200":
          description: Authenticated session
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/mfa/setup:
    post:
      operationId: setupMfa
      tags: [auth]
      summary: Start TOTP MFA enrollment
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        "200":
          description: Time-limited MFA provisioning details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaSetup"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/mfa/confirm:
    post:
      operationId: confirmMfa
      tags: [auth]
      summary: Confirm TOTP MFA enrollment
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaConfirmationInput"
      responses:
        "200":
          description: MFA enabled and recovery codes issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaConfirmation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/recovery:
    post:
      operationId: recoverAuth
      tags: [auth]
      summary: Authenticate using a recovery code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryInput"
      responses:
        "200":
          description: Authenticated session
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/password-reset:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a temporary password for a CMS user
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCredentialUpdate"
      responses:
        "200":
          description: User password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/password-change:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "204":
          description: Password changed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        "204":
          description: Session ended
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get the current session
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/csrf:
    get:
      operationId: getCsrfToken
      tags: [auth]
      summary: Issue a CSRF token for the current session
      security:
        - cookieAuth: []
      responses:
        "200":
          description: CSRF token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsrfToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cognirise_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    Conflict:
      description: Resource state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        details:
          type: object
          additionalProperties: true
...
    AuthBootstrap:
      type: object
      required: [setupRequired]
      properties:
        setupRequired: { type: boolean }
        csrfToken: { type: ["string", "null"] }
        session:
          anyOf:
            - $ref: "#/components/schemas/Session"
            - type: "null"
    AuthBootstrapInput:
      type: object
      additionalProperties: false
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$', maxLength: 254 }
        password: { type: string, minLength: 12, maxLength: 128 }
        mfaCode: { type: string, pattern: '^[0-9]{6}$' }
    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$', maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 128 }
    MfaVerificationInput:
      type: object
      additionalProperties: false
      required: [challengeId, code]
      properties:
        challengeId: { type: string }
        code: { type: string, pattern: '^[0-9]{6}$' }
    MfaSetup:
      type: object
      required: [secret, provisioningUri, expiresAt]
      properties:
        secret:
          type: string
          pattern: '^[A-Z2-7]+$'
          description: Base32-encoded TOTP secret. Display once and do not persist client-side.
        provisioningUri:
          type: string
          description: otpauth provisioning URI for a TOTP authenticator.
        expiresAt: { type: string, format: date-time }
    MfaConfirmationInput:
      type: object
      additionalProperties: false
      required: [secret, code]
      properties:
        secret:
          type: string
          pattern: '^[A-Z2-7]+$'
        code: { type: string, pattern: '^[0-9]{6}$' }
    MfaConfirmation:
      allOf:
        - $ref: "#/components/schemas/AuthResult"
        - type: object
          required: [recoveryCodes]
          properties:
            recoveryCodes:
              type: array
              minItems: 1
              items:
                type: string
                pattern: '^[A-Z0-9]{4,}(?:-[A-Z0-9]{4,})*$'
    RecoveryInput:
      type: object
      additionalProperties: false
      required: [email, recoveryCode]
      properties:
        email: { type: string, pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$', maxLength: 254 }
        recoveryCode:
          type: string
          minLength: 8
          maxLength: 64
    AuthResult:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        session:
          anyOf:
            - $ref: "#/components/schemas/Session"
            - type: "null"
        mfaChallenge:
          anyOf:
            - $ref: "#/components/schemas/MfaChallenge"
            - type: "null"
        csrfToken: { type: ["string", "null"] }
    MfaChallenge:
      type: object
      required: [id, expiresAt]
      properties:
        id: { type: string }
        method: { type: string, enum: [totp] }
        expiresAt: { type: string, format: date-time }
    Session:
      type: object
      required: [id, user, expiresAt]
      properties:
        id: { type: string }
        user:
          $ref: "#/components/schemas/User"
        mfaVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    CsrfToken:
      type: object
      required: [token, expiresAt]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time }
...
    UserRole:
      type: string
      enum: [administrator, publisher, editor, viewer]
    UserStatus:
      type: string
      enum: [invited, active, suspended]
    User:
      type: object
      required: [id, name, email, role, status, mfaEnabled, mustRotate, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$' }
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        mfaEnabled: { type: boolean }
        mustRotate: { type: boolean }
        lastLoginAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    UserCredentialUpdate:
      type: object
      additionalProperties: false
      required: [userId, temporaryPassword]
      properties:
        userId: { type: string }
        temporaryPassword:
          type: string
          minLength: 12
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[^A-Za-z0-9]).+$'
    PasswordChangeInput:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
          maxLength: 128
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
...