# lib/api-zod/src/generated/types/authBootstrap.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * First-party Cognirise website and mini CMS API.
 * OpenAPI spec version: 0.2.0
 */
import type { Session } from './session';

export interface AuthBootstrap {
  setupRequired: boolean;
  /** @nullable */
  csrfToken?: string | null;
  session?: Session | null;
}