# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * First-party Cognirise website and mini CMS API.
 * OpenAPI spec version: 0.2.0
 */
import type { MfaChallenge } from './mfaChallenge';
import type { Session } from './session';

export interface AuthResult {
  authenticated: boolean;
  session?: Session | null;
  mfaChallenge?: MfaChallenge | null;
  /** @nullable */
  csrfToken?: string | null;
}