# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * First-party Cognirise website and mini CMS API.
 * OpenAPI spec version: 0.2.0
 */

export interface PasswordChangeInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  currentPassword: string;
  /**
     * @minLength 12
     * @maxLength 128
     */
  newPassword: string;
}