# lib/api-zod/src/generated/types/userStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * First-party Cognirise website and mini CMS API.
 * OpenAPI spec version: 0.2.0
 */

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  invited: 'invited',
  active: 'active',
  suspended: 'suspended',
} as const;